tools/monorepo-tools/pyproject.toml

[project]
name = "via-monotools"
version = "0.3.1"
description = ""
dependencies = [
    "click>=8.1.3",
    "python-benedict[toml]>=0.30.2",
    "rich>=13.4.2",
    "tomlkit>=0.11.8",
]
requires-python = ">=3.10"
readme = "README.md"

[project.scripts]
monotools = "via_monotools.main:cli"
m_pylint = "via_monotools.linting:cli_pylint"
m_black = "via_monotools.linting:cli_black"
m_mypy = "via_monotools.linting:cli_mypy"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm.dev-dependencies]
dev = [
    "pylint>=2.17.4",
    "black>=23.3.0",
    "mypy>=1.3.0",
    "ruff>=0.0.277",
]

[tool.pdm.scripts]
format = {composite = [
    "ruff check --fix --exit-zero src/",
    "m_black src/",
]}
post_lock = "make"

lint = {composite = [
    "m_black --check src/",
    "ruff src/",
    "m_pylint src/",
    "m_mypy src/",
]}
